refactor: remove redundant PlanFormData interface
- Removed unnecessary wrapper interface
- Updated PlanForm component to use GeneratePlanFormData directly
