+++ app/api/generate/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { Configuration, OpenAIApi } from 'openai-edge';
-
+import { Recipe } from '../../types/mealPlanner';
 const config = new Configuration({
   apiKey: process.env.OPENAI_API_KEY
 });
