--- app/types/planFormData.ts
@@ -1,5 +1,0 @@
-import { CookingPrefs } from './cookingPrefs';
-import { GeneratePlanFormData } from './generatePlanForm';
-
-export interface PlanFormData extends GeneratePlanFormData, CookingPrefs {}
-
